+++ plone/app/upgrade/bbb.py
@@ -12,17 +12,6 @@ UndoTool = SimpleItem
 TinyMCE = SimpleItem
 
 
-class LanguageTool(SimpleItem):
-
-    def __call__(self, container, req):
-        pass
-
-    def manage_beforeDelete(self, item, container):
-        if item is self:
-            handle = self.meta_type + '/' + self.getId()
-            BeforeTraverse.unregisterBeforeTraverse(container, handle)
-
-
 class ILanguageTool(Interface):
     pass
