+++ plone/app/upgrade/bbbd/__init__.py
@@ -1,0 +1,2 @@
+from .LanguageTool import LanguageTool
+
